Move build-scripts/generate-from-templates to ssg/_build_templates.py
