+++ ssg/__init__.py
@@ -4,6 +4,7 @@ import ssg._build_guides as build_guides
 import ssg._build_renumber as build_renumber
 import ssg._build_roles as build_roles
 import ssg._build_stig as build_stig
+import ssg._build_templates as build_templates
 import ssg._build_yaml as build_yaml
 import ssg._checks as checks
 import ssg._constants as constants
